CozyBot/requirements.txt
pip==25.2

discord.py==2.6.3
psycopg2==2.9.10
SQLAlchemy==2.0.43
